addmoneymethod

+++ Project unit-6/PaymentWalletApplication/src/main/java/com/masai/entities/Wallet.java
@@ -33,7 +33,7 @@ public class Wallet {
 	
 	@Min(value = 0, message="Balance Should be greater than zero")
 	@NotNull
-	private BigDecimal balance;
+	private double balance;
 	
 	@OneToMany
 	private List<BankAccount> bankAccount = new ArrayList<>();
